requirements.txt
aioredis==1.3.1
asgiref==3.4.1
astroid==2.3.3
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2019.11.28
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
chardet==3.0.4
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
Django==3.2.7
django-ckeditor==6.1.0
django-colorfield==0.5.0
django-crispy-forms==1.13.0
django-js-asset==1.2.2
django-multiselectfield==0.1.12
hiredis==2.0.0
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.2
Pillow==8.3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==21.0.0
pytz==2021.1
requests==2.22.0
service-identity==21.1.0
six==1.13.0
sqlparse==0.4.2
sslcommerz-python==0.0.7
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typed-ast==1.4.0
typing-extensions==3.10.0.2
urllib3==1.25.7
wrapt==1.11.2
zope.interface==5.4.0
